提出前の全部のプログラム

+++ app/Http/Controllers/ManagementController.php
@@ -5,11 +5,13 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Money_management;
 use Illuminate\Support\Facades\Auth;
+use App\Services\ManagementService;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 
 class ManagementController extends Controller
 {
-        /**
+    /**
      * Handle the incoming request.
      *
      * @param  \Illuminate\Http\Request  $request
@@ -18,7 +20,13 @@ class ManagementController extends Controller
     public function show(Request $request)
     {
         $managements = Money_management::orderBy('created_at', 'DESC')->get();
-        return view('management.index')->with('managements', $managements);
+
+        $total = 0;
+        foreach($managements as $management) {
+            $total += $management->amount_money;
+        }
+
+        return view('management.index')->with('managements', $managements)->with('total', $total);
     }
 
     /**
@@ -52,9 +60,14 @@ class ManagementController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function updateIndex(Request $request)
+    public function updateIndex(Request $request, ManagementService $managementService)
     {
         $managementId = (int) $request->route('managementId');
+
+        if(!$managementService->checkOwnManagement(Auth::id(), $managementId)) {
+            throw new AccessDeniedHttpException;
+        }
+
         $management = Money_management::where('id', $managementId)->firstOrFail();
         return view('management.update')->with('management', $management);
     }
@@ -65,8 +78,13 @@ class ManagementController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function updatePut(Request $request)
+    public function updatePut(Request $request, ManagementService $managementService)
     {
+
+        if(!$managementService->checkOwnManagement(Auth::id(), $request->managementId)) {
+            throw new AccessDeniedHttpException;
+        }
+
         $validatedData = $request->validate([
             'purchase'      => 'required',
             'date'          => 'required',
@@ -89,9 +107,14 @@ class ManagementController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function delete(Request $request)
+    public function delete(Request $request, ManagementService $managementService)
     {
         $managementId = (int) $request->route('managementId');
+
+        if(!$managementService->checkOwnManagement(Auth::id(), $request->managementId)) {
+            throw new AccessDeniedHttpException;
+        }
+
         $management = Money_management::where('id', $managementId)->firstOrFail();
         $management->delete();
         return redirect()->route('management.show')->with('feedback.success', "家計簿を削除しました");
